Security configuration refactory
