+++ iam-login-service/src/main/java/it/infn/mw/iam/config/SecurityConfig.java
@@ -1,34 +1,30 @@
 /**
  * Copyright (c) Istituto Nazionale di Fisica Nucleare (INFN). 2016-2018
  *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License. You may obtain a copy of the License at
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
  *
- * http://www.apache.org/licenses/LICENSE-2.0
+ *     http://www.apache.org/licenses/LICENSE-2.0
  *
- * Unless required by applicable law or agreed to in writing, software distributed under the License
- * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
- * or implied. See the License for the specific language governing permissions and limitations under
- * the License.
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 package it.infn.mw.iam.config;
-
-import static it.infn.mw.iam.api.tokens.Constants.ACCESS_TOKENS_ENDPOINT;
-import static it.infn.mw.iam.api.tokens.Constants.REFRESH_TOKENS_ENDPOINT;
 
 import org.mitre.oauth2.service.OAuth2TokenEntityService;
 import org.mitre.oauth2.web.CorsFilter;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.context.embedded.FilterRegistrationBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
 import org.springframework.core.Ordered;
 import org.springframework.core.annotation.Order;
-import org.springframework.http.HttpMethod;
-import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.builders.WebSecurity;
@@ -36,216 +32,18 @@ import org.springframework.security.config.annotation.web.configuration.EnableWe
 import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
 import org.springframework.security.config.http.SessionCreationPolicy;
 import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.oauth2.provider.authentication.OAuth2AuthenticationManager;
 import org.springframework.security.oauth2.provider.authentication.OAuth2AuthenticationProcessingFilter;
 import org.springframework.security.oauth2.provider.client.ClientCredentialsTokenEndpointFilter;
 import org.springframework.security.oauth2.provider.error.OAuth2AuthenticationEntryPoint;
-import org.springframework.security.oauth2.provider.expression.OAuth2WebSecurityExpressionHandler;
-import org.springframework.security.web.access.AccessDeniedHandlerImpl;
-import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
 import org.springframework.security.web.authentication.Http403ForbiddenEntryPoint;
-import org.springframework.security.web.authentication.LoginUrlAuthenticationEntryPoint;
 import org.springframework.security.web.authentication.www.BasicAuthenticationFilter;
 import org.springframework.security.web.context.SecurityContextPersistenceFilter;
 import org.springframework.security.web.context.request.async.WebAsyncManagerIntegrationFilter;
-import org.springframework.security.web.savedrequest.HttpSessionRequestCache;
-import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
-import org.springframework.web.filter.GenericFilterBean;
-
-import it.infn.mw.iam.api.account.AccountUtils;
-import it.infn.mw.iam.api.aup.AUPSignatureCheckService;
-import it.infn.mw.iam.authn.EnforceAupSignatureSuccessHandler;
-import it.infn.mw.iam.authn.RootIsDashboardSuccessHandler;
-import it.infn.mw.iam.authn.oidc.OidcAccessDeniedHandler;
-import it.infn.mw.iam.authn.oidc.OidcAuthenticationProvider;
-import it.infn.mw.iam.authn.oidc.OidcClientFilter;
-import it.infn.mw.iam.authn.x509.IamX509AuthenticationProvider;
-import it.infn.mw.iam.authn.x509.IamX509AuthenticationUserDetailService;
-import it.infn.mw.iam.authn.x509.IamX509PreauthenticationProcessingFilter;
-import it.infn.mw.iam.authn.x509.X509AuthenticationCredentialExtractor;
-import it.infn.mw.iam.persistence.repository.IamAccountRepository;
 
 @Configuration
 @EnableWebSecurity
 public class SecurityConfig {
-
-  @Configuration
-  @Order(100)
-  public static class UserLoginConfig extends WebSecurityConfigurerAdapter {
-
-    @Value("${iam.baseUrl}")
-    private String iamBaseUrl;
-
-    @Autowired
-    private OAuth2WebSecurityExpressionHandler oAuth2WebSecurityExpressionHandler;
-
-    @Autowired
-    @Qualifier("mitreAuthzRequestFilter")
-    private GenericFilterBean authorizationRequestFilter;
-
-    @Autowired
-    @Qualifier("iamUserDetailsService")
-    private UserDetailsService iamUserDetailsService;
-
-    @Autowired
-    private X509AuthenticationCredentialExtractor x509CredentialExtractor;
-
-    @Autowired
-    private IamX509AuthenticationUserDetailService x509UserDetailsService;
-
-    @Autowired
-    private PasswordEncoder passwordEncoder;
-
-    @Autowired
-    private IamAccountRepository accountRepo;
-
-    @Autowired
-    private AUPSignatureCheckService aupSignatureCheckService;
-
-    @Autowired
-    private AccountUtils accountUtils;
-
-    @Autowired
-    public void configureGlobal(final AuthenticationManagerBuilder auth) throws Exception {
-      // @formatter:off
-      auth
-        .userDetailsService(iamUserDetailsService)
-        .passwordEncoder(passwordEncoder);
-      // @formatter:on
-    }
-
-    @Override
-    public void configure(final WebSecurity web) throws Exception {
-
-      web.expressionHandler(oAuth2WebSecurityExpressionHandler);
-    }
-
-
-    public IamX509AuthenticationProvider iamX509AuthenticationProvider() {
-
-      IamX509AuthenticationProvider provider = new IamX509AuthenticationProvider();
-      provider.setPreAuthenticatedUserDetailsService(x509UserDetailsService);
-      return provider;
-    }
-
-
-    public IamX509PreauthenticationProcessingFilter iamX509Filter() throws Exception {
-      return new IamX509PreauthenticationProcessingFilter(x509CredentialExtractor,
-          iamX509AuthenticationProvider(), successHandler());
-
-    }
-
-    @Override
-    protected void configure(final HttpSecurity http) throws Exception {
-
-      // @formatter:off
-
-      http.requestMatchers()
-        .antMatchers("/", "/login**", "/logout", "/authorize", "/manage/**", "/dashboard**", "/register",
-            "/reset-session", "/device/**")
-        .and()
-        .sessionManagement()
-          .enableSessionUrlRewriting(false)
-        .and()
-          .authorizeRequests()
-            .antMatchers("/login**", "/webjars/**").permitAll()
-            .antMatchers("/register").permitAll()
-            .antMatchers("/authorize**").permitAll()
-            .antMatchers("/reset-session").permitAll()
-            .antMatchers("/device/**").authenticated()
-            .antMatchers("/").authenticated()
-        .and()
-          .formLogin()
-            .loginPage("/login")
-            .failureUrl("/login?error=failure")
-            .successHandler(successHandler())
-        .and()
-          .exceptionHandling()
-            .accessDeniedHandler(new OidcAccessDeniedHandler())
-        .and()
-          .addFilterBefore(authorizationRequestFilter, SecurityContextPersistenceFilter.class)
-        .logout()
-          .logoutUrl("/logout")
-        .and().anonymous()
-        .and()
-          .csrf()
-            .requireCsrfProtectionMatcher(new AntPathRequestMatcher("/authorize")).disable()
-        .addFilter(iamX509Filter());
-       
-      // @formatter:on
-    }
-
-    @Bean
-    public OAuth2WebSecurityExpressionHandler oAuth2WebSecurityExpressionHandler() {
-
-      return new OAuth2WebSecurityExpressionHandler();
-    }
-
-    public AuthenticationSuccessHandler successHandler() {
-      AuthenticationSuccessHandler delegate =
-          new RootIsDashboardSuccessHandler(iamBaseUrl, new HttpSessionRequestCache());
-
-
-      return new EnforceAupSignatureSuccessHandler(delegate, aupSignatureCheckService, accountUtils,
-          accountRepo);
-    }
-  }
-
-  @Configuration
-  @Order(105)
-  @Profile("google")
-  public static class ExternalOidcLogin extends WebSecurityConfigurerAdapter {
-
-    @Autowired
-    @Qualifier("OIDCAuthenticationManager")
-    private AuthenticationManager oidcAuthManager;
-
-    @Autowired
-    OidcAuthenticationProvider authProvider;
-
-    @Autowired
-    @Qualifier("OIDCAuthenticationFilter")
-    private OidcClientFilter oidcFilter;
-
-    @Override
-    public AuthenticationManager authenticationManagerBean() throws Exception {
-
-      return oidcAuthManager;
-    }
-
-    // @Bean(name = "ExternalAuthenticationEntryPoint")
-    public LoginUrlAuthenticationEntryPoint authenticationEntryPoint() {
-
-      return new LoginUrlAuthenticationEntryPoint("/openid_connect_login");
-    }
-
-    @Override
-    protected void configure(final AuthenticationManagerBuilder auth) throws Exception {
-
-      auth.authenticationProvider(authProvider);
-    }
-
-    @Override
-    protected void configure(final HttpSecurity http) throws Exception {
-
-      // @formatter:off
-      http
-        .antMatcher("/openid_connect_login**")
-          .exceptionHandling()
-            .authenticationEntryPoint(authenticationEntryPoint())
-          .accessDeniedHandler(new OidcAccessDeniedHandler())
-        .and()
-          .addFilterAfter(oidcFilter, SecurityContextPersistenceFilter.class).authorizeRequests()
-        .antMatchers("/openid_connect_login**")
-          .permitAll()
-        .and()
-          .sessionManagement()
-            .enableSessionUrlRewriting(false)
-            .sessionCreationPolicy(SessionCreationPolicy.ALWAYS);
-      // @formatter:on
-    }
-  }
 
   @Configuration
   @Order(9)
@@ -396,7 +194,6 @@ public class SecurityConfig {
     private CorsFilter corsFilter;
 
     @Override
-
     public void configure(final HttpSecurity http) throws Exception {
 
       // @formatter:off
@@ -530,189 +327,6 @@ public class SecurityConfig {
   }
 
   @Configuration
-  @Order(18)
-  public static class IamApiConfig extends WebSecurityConfigurerAdapter {
-
-    private static final String AUP_PATH = "/iam/aup";
-    private static final String GROUP_REQUEST_PATH = "/iam/group_requests";
-
-    @Autowired
-    private OAuth2AuthenticationProcessingFilter resourceFilter;
-
-    @Autowired
-    private OAuth2AuthenticationEntryPoint authenticationEntryPoint;
-
-    @Autowired
-    private CorsFilter corsFilter;
-
-    @Override
-    protected void configure(final HttpSecurity http) throws Exception {
-      // @formatter:off
-      http.requestMatchers()
-        .antMatchers("/scim/**", 
-            "/registration/**", 
-            "/iam/password-update", 
-            "/iam/scope_policies/**",
-            ACCESS_TOKENS_ENDPOINT + "/**", REFRESH_TOKENS_ENDPOINT + "/**",
-            "/iam/aup/**",
-            GROUP_REQUEST_PATH+"**", GROUP_REQUEST_PATH+"/**",
-            "/iam/account/search**", 
-            "/iam/group/search**")
-        .and()
-          .exceptionHandling()
-            .authenticationEntryPoint(authenticationEntryPoint)
-        .and()
-          .addFilterAfter(resourceFilter, SecurityContextPersistenceFilter.class)
-          .addFilterBefore(corsFilter, WebAsyncManagerIntegrationFilter.class)
-        .sessionManagement()
-          .sessionCreationPolicy(SessionCreationPolicy.NEVER)
-        .and()
-          .authorizeRequests()
-            .antMatchers("/scim/**").authenticated()
-            .antMatchers(HttpMethod.POST, "/registration/create").permitAll()
-            .antMatchers(HttpMethod.GET, "/registration/username-available/**").permitAll()
-            .antMatchers(HttpMethod.GET, "/registration/email-available/**").permitAll()
-            .antMatchers(HttpMethod.GET, "/registration/confirm/**").permitAll()
-            .antMatchers(HttpMethod.GET, "/registration/verify/**").permitAll()
-            .antMatchers(HttpMethod.GET, "/registration/submitted").permitAll()
-            .antMatchers("/registration/**").authenticated()
-            .antMatchers("/iam/password-update").authenticated()
-            .antMatchers("/iam/scope_policies/**").authenticated()
-            .antMatchers(ACCESS_TOKENS_ENDPOINT + "/**", REFRESH_TOKENS_ENDPOINT + "/**").authenticated()
-            .antMatchers(HttpMethod.GET, AUP_PATH).permitAll()
-            .antMatchers(HttpMethod.POST, AUP_PATH).authenticated()
-            .antMatchers(HttpMethod.DELETE, AUP_PATH).authenticated()
-            .antMatchers(GROUP_REQUEST_PATH+"**", GROUP_REQUEST_PATH+"/**" ).authenticated()
-            .antMatchers(HttpMethod.GET, "/iam/account/search", "/iam/group/search").permitAll()
-        .and()
-          .csrf().disable();
-      // @formatter:on
-    }
-  }
-
-
-  @Configuration
-  @Order(20)
-  public static class AuthnInfoEndpointConfig extends WebSecurityConfigurerAdapter {
-    @Autowired
-    private OAuth2AuthenticationEntryPoint authenticationEntryPoint;
-
-    @Override
-    protected void configure(HttpSecurity http) throws Exception {
-      http.antMatcher("/iam/authn-info/**")
-        .exceptionHandling()
-        .authenticationEntryPoint(authenticationEntryPoint)
-        .and()
-        .sessionManagement()
-        .sessionCreationPolicy(SessionCreationPolicy.NEVER)
-        .and()
-        .authorizeRequests()
-        .antMatchers("/iam/authn-info/**")
-        .authenticated();
-    }
-  }
-
-  @Configuration
-  @Order(21)
-  public static class AccountLinkingEndpointConfig extends WebSecurityConfigurerAdapter {
-    @Autowired
-    private OAuth2AuthenticationEntryPoint authenticationEntryPoint;
-
-    @Override
-    protected void configure(HttpSecurity http) throws Exception {
-      http.antMatcher("/iam/account-linking/**")
-        .exceptionHandling()
-        .authenticationEntryPoint(authenticationEntryPoint)
-        .and()
-        .sessionManagement()
-        .sessionCreationPolicy(SessionCreationPolicy.NEVER)
-        .and()
-        .csrf()
-        .disable()
-        .authorizeRequests()
-        .antMatchers("/iam/account-linking/**")
-        .authenticated();
-    }
-  }
-
-  @Configuration
-  @Order(22)
-  public static class AccountAuthorityEndpointConfig extends WebSecurityConfigurerAdapter {
-    @Autowired
-    private OAuth2AuthenticationEntryPoint authenticationEntryPoint;
-
-    @Override
-    protected void configure(HttpSecurity http) throws Exception {
-      http.requestMatchers()
-        .antMatchers("/iam/account/**", "/iam/me/**")
-        .and()
-        .exceptionHandling()
-        .authenticationEntryPoint(authenticationEntryPoint)
-        .and()
-        .sessionManagement()
-        .sessionCreationPolicy(SessionCreationPolicy.NEVER)
-        .and()
-        .authorizeRequests()
-        .antMatchers("/iam/account/**", "/iam/me/**")
-        .authenticated()
-        .and()
-        .csrf()
-        .disable();
-    }
-  }
-
-  @Configuration
-  @Order(24)
-  public static class ActuatorEndpointsConfig extends WebSecurityConfigurerAdapter {
-
-    @Value("${iam.superuser.username}")
-    private String basicUsername;
-
-    @Value("${iam.superuser.password}")
-    private String basicPassword;
-
-    @Autowired
-    private OAuth2AuthenticationProcessingFilter resourceFilter;
-
-    @Autowired
-    private CustomAuthenticationEntryPoint customAuthenticationEntyPoint;
-
-    @Override
-    public void configure(AuthenticationManagerBuilder auth) throws Exception {
-      // @formatter:off
-      auth.inMemoryAuthentication()
-        .withUser(basicUsername).password(basicPassword)
-        .roles("SUPERUSER", "ADMIN");
-      // @formatter:on
-    }
-
-    @Override
-    protected void configure(final HttpSecurity http) throws Exception {
-      // @formatter:off
-      http
-        .requestMatchers()
-          .antMatchers("/metrics", "/info", "/health", "/health/mail", "/health/external",
-              "/configprops", "/env", "/mappings", "/flyway", "/autoconfig", "/beans", "/dump", "/trace")
-        .and()
-          .httpBasic()
-          .authenticationEntryPoint(customAuthenticationEntyPoint)
-        .and()
-          .exceptionHandling()
-            .accessDeniedHandler(new AccessDeniedHandlerImpl())
-        .and()
-          .addFilterAfter(resourceFilter, SecurityContextPersistenceFilter.class)
-        .sessionManagement()
-          .sessionCreationPolicy(SessionCreationPolicy.NEVER)
-        .and()
-          .authorizeRequests()
-            .antMatchers(HttpMethod.GET, "/info", "/health", "/health/mail", "/health/external").permitAll()
-            .antMatchers(HttpMethod.GET, "/metrics").hasRole("ADMIN")
-            .antMatchers(HttpMethod.GET, "/configprops", "/env", "/mappings", "/flyway", "/autoconfig", "/beans", "/dump", "/trace").hasRole("SUPERUSER");
-      // @formatter:on
-    }
-  }
-
-  @Configuration
   @Order(27)
   public static class DeviceCodeEnpointConfig extends WebSecurityConfigurerAdapter {
 
@@ -774,7 +388,6 @@ public class SecurityConfig {
 
     @Override
     public void configure(final WebSecurity builder) throws Exception {
-
       builder.debug(true);
     }
   }
